refactor: rename `initialization.js` => `state.js`
