+++ server/src/models/Backup.js
@@ -22,7 +22,7 @@ class Backup extends Model {
   }
 
   static get relationMappings() {
-    const { Db } = require('../services/initialization')
+    const { Db } = require('../services/state')
     return {
       user: {
         relation: Model.BelongsToOneRelation,
