+++ server/src/routes/rootRouter.js
@@ -4,7 +4,7 @@ const { join } = require('path')
 const { SourceMapConsumer } = require('source-map')
 const config = require('@rm/config')
 const { log, HELPERS } = require('@rm/logger')
-const { Event, Db } = require('../services/initialization')
+const { Event, Db } = require('../services/state')
 const { version } = require('../../../package.json')
 const areaPerms = require('../services/functions/areaPerms')
 const getServerSettings = require('../services/functions/getServerSettings')
