+++ server/src/services/events/guildMemberRemove.js
@@ -1,6 +1,6 @@
 // @ts-check
 const { log, HELPERS } = require('@rm/logger')
-const { Db } = require('../initialization')
+const { Db } = require('../state')
 
 /**
  *
